优化对象管理用例
git-svn-id: http://192.168.20.11/sequoiadb/trunk@38008 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ testcase_new/s3/java/src/test/java/com/sequoias3/object/GetObjectWithRange16355.java
@@ -56,8 +56,11 @@ public class GetObjectWithRange16355 extends S3TestBase{
 		TestTools.LocalFile.removeFile(localPath);
 		TestTools.LocalFile.createDir(localPath.toString());
 		TestTools.LocalFile.createFile(filePath, fileSize);		
-		s3Client = CommLib.buildS3Client();			
-		s3Client.deleteObject(S3TestBase.bucketName, key);
+		s3Client = CommLib.buildS3Client();	
+		
+		if( s3Client.doesObjectExist(S3TestBase.bucketName, key)){			
+    		s3Client.deleteObject(S3TestBase.bucketName,key);
+    	}		
 	}
 
 	@Test(dataProvider = "rangeProvider")
@@ -71,7 +74,7 @@ public class GetObjectWithRange16355 extends S3TestBase{
 	private void tearDown() {
 		try {
 			if (runSuccess) {
-				
+				s3Client.deleteObject(S3TestBase.bucketName,key);
 			}
 		} finally {
 		    s3Client.shutdown();
